# PAWProjekt/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Flask REST API",
        "description": "API documentation",
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/users": {
            "get": {
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "A list of users"
                    }
                }
            },
            "post": {
                "summary": "Create a new user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A user object"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "summary": "Update a user by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a user by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "summary": "Get all posts",
                "responses": {
                    "200": {
                        "description": "A list of posts"
                    }
                }
            },
            "post": {
                "summary": "Create a new post",
                "parameters": [
                    {
                        "name": "post",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post created"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "summary": "Get a post by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A post object"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "put": {
                "summary": "Update a post by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "post",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post updated"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a post by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        },
        "/comments": {
            "get": {
                "summary": "Get all comments",
                "responses": {
                    "200": {
                        "description": "A list of comments"
                    }
                }
            },
            "post": {
                "summary": "Create a new comment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comment created"
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "summary": "Get a comment by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A comment object"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            },
            "put": {
                "summary": "Update a comment by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "comment",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment updated"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a comment by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted"
                    },
                    "404": {
                        "description": "Comment not found"
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "summary": "Upload a file",
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "File uploaded"
                    },
                    "400": {
                        "description": "No file part or no selected file"
                    }
                }
            }
        },
        "/download/{filename}": {
            "get": {
                "summary": "Download a file",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded"
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login a user",
                "parameters": [
                    {
                        "name": "credentials",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "access_token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/protected": {
            "get": {
                "summary": "Access protected route",
                "responses": {
                    "200": {
                        "description": "Access granted"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Post": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "Comment": {
            "type": "object",
            "properties": {
                "post_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}